// src/lib/api/generated/model/emailListStatusEnum.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ACTIVE` - Active
 * `INACTIVE` - Inactive
 */
export type EmailListStatusEnum = (typeof EmailListStatusEnum)[keyof typeof EmailListStatusEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EmailListStatusEnum = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const;